zpay32: move the BOLT-11 compatible invoice format to zpay32 namespace
