+++ zpay32/amountunits.go
@@ -1,4 +1,4 @@
-package invoice
+package zpay32
 
 import (
 	"fmt"
