+++ zpay32/invoice.go
@@ -1,4 +1,4 @@
-package invoice
+package zpay32
 
 import (
 	"bytes"
